Ordered future sequencing
The current Future.sequence method does not maintain the order of the
input Future collection, as it inserts the Future results into the final
collection as the Futures resolve, and there is no guarantee they will
resolve in the same order they are passed into the method.

Since most Scala methods that operate on collections maintain the
ordering of the collection, it can be surprising to find that this one
does not.

This commit introduces sequenceOrdered, which guarantees the ordering of
the result collection to match that of the input collection. It does
this by storing the result of the Future zipped with its original index
in a TreeSet, ensuring that the later extraction of the results happens
in the same order.